feat: refactor out RestoreFilePipeline
