+++ src/Arius.Core/Features/RestoreCommand/RestoreCommandHandler.cs
@@ -66,9 +66,7 @@ public sealed class RestoreCommandHandler(
 {
     // ── Concurrency knobs ──────────────────────────────────────────────────────
 
-    private const int RouteWorkers     = 8;
-    private const int ResolveBatchSize = 32;
-    private const int DownloadWorkers  = 4;
+    private const int DownloadWorkers = 4;
 
     /// <summary>
     /// Executes the end-to-end restore pipeline for the provided <see cref="RestoreCommand"/>. See the
@@ -83,6 +81,7 @@ public sealed class RestoreCommandHandler(
         try
         {
             var fs = new RelativeFileSystem(LocalDirectory.Parse(opts.RootDirectory));
+            var filePipeline = new RestoreFilePipeline(encryption, index, fileTreeService, mediator, logger);
 
             // ── Stage 1: Resolve snapshot ─────────────────────────────────────────
             logger.LogInformation("[phase] resolve-snapshot");
@@ -112,7 +111,7 @@ public sealed class RestoreCommandHandler(
             var classification  = new Dictionary<ChunkHash, ChunkClassification>();
             var fileCount       = 0;
 
-            await foreach (var resolved in StreamResolvedFilesAsync(fs, snapshot.RootHash, opts, emitEvents: true, skipped, cancellationToken))
+            await foreach (var resolved in filePipeline.StreamResolvedFilesAsync(fs, snapshot.RootHash, opts, emitEvents: true, skipped, cancellationToken))
             {
                 fileCount++;
                 var entry     = resolved.IndexEntry;
@@ -226,7 +225,7 @@ public sealed class RestoreCommandHandler(
                     var openTars = new Dictionary<ChunkHash, List<FileToRestore>>();
                     try
                     {
-                        await foreach (var resolved in StreamResolvedFilesAsync(fs, snapshot.RootHash, opts, emitEvents: false, skipped: null, ct))
+                        await foreach (var resolved in filePipeline.StreamResolvedFilesAsync(fs, snapshot.RootHash, opts, emitEvents: false, skipped: null, ct))
                         {
                             var chunkHash = resolved.IndexEntry.ChunkHash;
                             if (!classification.TryGetValue(chunkHash, out var cc) || cc.Status != ChunkHydrationStatus.Available)
@@ -394,174 +393,6 @@ public sealed class RestoreCommandHandler(
             };
         }
     }
-
-
-    // ── Pipeline: Walk → Route → Resolve used for Stage 2 (Classify) & Stage 4 (Download) composed as one stream ─────────────────
-
-    /// <summary>
-    /// All files from the selected snapshot/target path that the restore command intends to bring back locally,
-    /// subject to overwrite/skip rules, paired with the chunk-index entry needed to restore them.
-    /// </summary>
-    private async IAsyncEnumerable<ResolvedFile> StreamResolvedFilesAsync(
-        RelativeFileSystem fs,
-        FileTreeHash       rootHash,
-        RestoreOptions     opts,
-        bool               emitEvents,
-        StrongBox<long>?   skipped,
-        [EnumeratorCancellation] CancellationToken cancellationToken)
-    {
-        var files = WalkAsync(rootHash, opts.TargetPath, emitProgress: emitEvents, cancellationToken)
-            .WhereParallelAsync(
-                RouteWorkers,
-                (file, ct) => ShouldRestoreAsync(file, fs, opts, emitEvents, skipped, ct),
-                cancellationToken);
-
-        await foreach (var batch in files.Chunk(ResolveBatchSize).WithCancellation(cancellationToken))
-            await foreach (var resolved in ResolveBatchAsync(batch, cancellationToken))
-                yield return resolved;
-    }
-
-    // ── Pipeline: Walk (breadth-first, like ListQueryHandler) ───────
-
-    /// <summary>
-    /// Yields one <see cref="FileToRestore"/> per remote file that matches <paramref name="targetPrefix"/>
-    /// (or all files when <c>null</c>) and emits restore-specific traversal progress.
-    /// </summary>
-    private async IAsyncEnumerable<FileToRestore> WalkAsync(
-        FileTreeHash  rootHash,
-        RelativePath? targetPrefix,
-        bool          emitProgress,
-        [EnumeratorCancellation] CancellationToken cancellationToken)
-    {
-        var total                 = 0;
-        var emittedSinceLastEvent = 0;
-        var lastEmit              = DateTimeOffset.UtcNow;
-
-        var walker = new FileTreeWalker(fileTreeService);
-        await foreach (var file in walker.WalkFilesAsync(rootHash, targetPrefix, cancellationToken).ConfigureAwait(false))
-        {
-            yield return file;
-            total++;
-            emittedSinceLastEvent++;
-
-            if (emitProgress)
-            {
-                var now = DateTimeOffset.UtcNow;
-                if (emittedSinceLastEvent >= 10 || (now - lastEmit).TotalMilliseconds >= 100)
-                {
-                    emittedSinceLastEvent = 0;
-                    lastEmit = now;
-                    logger.LogDebug("[tree] Traversal progress: {FilesFound} files discovered", total);
-                    await mediator.Publish(new TreeTraversalProgressEvent(total), cancellationToken);
-                }
-            }
-        }
-
-        if (emitProgress && total > 0)
-            await mediator.Publish(new TreeTraversalProgressEvent(total), cancellationToken);
-    }
-
-    // ── Pipeline: Route (conflict check) ────────────────────────────
-
-    /// <summary>
-    /// Decides one file's fate against the local filesystem and returns whether it should be restored:
-    /// skip identical files, keep locally-differing files unless <c>--overwrite</c>, otherwise restore.
-    /// Per-file events are emitted only when <paramref name="emitEvents"/> is set (classify pass).
-    /// Runs on multiple route workers, so <paramref name="skipped"/> is updated with <see cref="Interlocked"/>.
-    /// </summary>
-    private async ValueTask<bool> ShouldRestoreAsync(
-        FileToRestore      file,
-        RelativeFileSystem fs,
-        RestoreOptions     opts,
-        bool               emitEvents,
-        StrongBox<long>?   skipped,
-        CancellationToken  cancellationToken)
-    {
-        if (fs.FileExists(file.RelativePath))
-        {
-            if (!opts.Overwrite)
-            {
-                // Hash the local file to check whether it already matches.
-                await using var s = fs.OpenRead(file.RelativePath);
-                var localHash = await encryption.ComputeHashAsync(s, cancellationToken);
-
-                if (localHash == file.ContentHash)
-                {
-                    if (skipped is not null) 
-                        Interlocked.Increment(ref skipped.Value);
-
-                    if (emitEvents)
-                    {
-                        logger.LogInformation("[route] {Path} -> skip (identical)", file.RelativePath);
-                        await mediator.Publish(new FileSkippedEvent(file.RelativePath, s.Length), cancellationToken);
-                        await mediator.Publish(new FileRoutedEvent(file.RelativePath, RestoreRoute.SkipIdentical, s.Length), cancellationToken);
-                    }
-
-                    return false;
-                }
-                else
-                {
-                    // Exists with a different hash and no --overwrite → keep the local copy.
-                    if (skipped is not null)
-                        Interlocked.Increment(ref skipped.Value);
-
-                    if (emitEvents)
-                    {
-                        logger.LogInformation("[route] {Path} -> keep (local differs, no --overwrite)", file.RelativePath);
-                        await mediator.Publish(new FileSkippedEvent(file.RelativePath, s.Length),                               cancellationToken);
-                        await mediator.Publish(new FileRoutedEvent(file.RelativePath, RestoreRoute.KeepLocalDiffers, s.Length), cancellationToken);
-                    }
-                    return false;
-                }
-            }
-
-            if (emitEvents)
-            {
-                logger.LogInformation("[route] {Path} -> overwrite", file.RelativePath);
-                await mediator.Publish(new FileRoutedEvent(file.RelativePath, RestoreRoute.Overwrite, fs.GetFileSize(file.RelativePath)), cancellationToken);
-            }
-        }
-        else if (emitEvents)
-        {
-            logger.LogInformation("[route] {Path} -> new", file.RelativePath);
-            await mediator.Publish(new FileRoutedEvent(file.RelativePath, RestoreRoute.New, 0), cancellationToken);
-        }
-
-        return true;
-    }
-
-    // ── Pipeline: Resolve (batched chunk-index lookup) ──────────────
-
-    private async IAsyncEnumerable<ResolvedFile> ResolveBatchAsync(
-        IReadOnlyList<FileToRestore> batch,
-        [EnumeratorCancellation] CancellationToken cancellationToken)
-    {
-        if (batch.Count == 0)
-            yield break;
-
-        var entries = await index.LookupAsync(batch.Select(f => f.ContentHash).Distinct(), cancellationToken);
-
-        List<FileToRestore>? missing = null;
-        foreach (var file in batch)
-        {
-            if (!entries.TryGetValue(file.ContentHash, out var entry))
-            {
-                (missing ??= []).Add(file);
-                continue;
-            }
-
-            yield return new ResolvedFile(file, entry);
-        }
-
-        if (missing is not null)
-        {
-            var sample = string.Join(", ", missing.Take(5).Select(f => $"{f.RelativePath} ({f.ContentHash.Short8})"));
-            throw new InvalidOperationException($"Snapshot references {missing.Count} content hash(es) that are missing from the chunk index: {sample}. Run the explicit chunk-index repair command and retry restore.");
-        }
-    }
-
-    // ── Pipeline: Walk → Route → Resolve END
-
 
     private static ChunkHydrationStatus ClassifyChunk(ChunkHash chunkHash, ShardEntry entry, IReadOnlyDictionary<ChunkHash, bool> rehydratedState)
     {
